Moved CoreData classes in subgroups Added custom getters/setters to convert Int64 to Int Added computed variable for representing lastEpisodeWatched and lastSeasonWatched Added persistentContainer code to AppDelegate Started modeling MediaLibrary in CoreData

+++ Movie DB/Model/CastMember/CastMember+CoreDataClass.swift
@@ -16,7 +16,7 @@ public class CastMember: NSManagedObject, Decodable {
     
     // MARK: - Initializers
     
-    public convenience init(context: NSManagedObjectContext, id: Int64, name: String, roleName: String, imagePath: String? = nil) {
+    public convenience init(context: NSManagedObjectContext, id: Int, name: String, roleName: String, imagePath: String? = nil) {
         self.init(context: context)
         self.id = id
         self.name = name
@@ -33,7 +33,7 @@ public class CastMember: NSManagedObject, Decodable {
         self.init(context: context)
         
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = try container.decode(Int64.self, forKey: .id)
+        self.id = try container.decode(Int.self, forKey: .id)
         self.name = try container.decode(String.self, forKey: .name)
         self.roleName = try container.decode(String.self, forKey: .roleName)
         self.imagePath = try container.decode(String?.self, forKey: .imagePath)
