+++ Movie DB/Model/Video/Video+CoreDataClass.swift
@@ -25,7 +25,7 @@ public class Video: NSManagedObject, Decodable {
         self.name = try container.decode(String.self, forKey: .name)
         self.site = try container.decode(String.self, forKey: .site)
         self.type = try container.decode(String.self, forKey: .type)
-        self.resolution = try container.decode(Int64.self, forKey: .resolution)
+        self.resolution = try container.decode(Int.self, forKey: .resolution)
         self.language = try container.decode(String.self, forKey: .language)
         self.region = try container.decode(String.self, forKey: .region)
     }
